fixed navbar and set active color for active route
